// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type BlogArticleDocumentDataSlicesSlice =
  | TextSlice
  | FormSlice
  | ArticleListSlice
  | HighlightedTextSlice
  | TestimonialsSlice
  | CallToActionSlice;

/**
 * Item in *Blog Article → Social Cards - Facebook & Twitter*
 */
export interface BlogArticleDocumentDataSocialCardsItem {
  /**
   * Social Card Image field in *Blog Article → Social Cards - Facebook & Twitter*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.social_cards[].social_card_image
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  social_card_image: prismic.ImageField<never>;

  /**
   * Social Card Title field in *Blog Article → Social Cards - Facebook & Twitter*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.social_cards[].social_card_title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  social_card_title: prismic.KeyTextField;

  /**
   * Social card description field in *Blog Article → Social Cards - Facebook & Twitter*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.social_cards[].social_card_description
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  social_card_description: prismic.KeyTextField;
}

/**
 * Content for Blog Article documents
 */
interface BlogArticleDocumentData {
  /**
   * Title field in *Blog Article*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  title: prismic.RichTextField;

  /**
   * Author field in *Blog Article*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.author
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  author: prismic.ContentRelationshipField<"author">;

  /**
   * Excerpt field in *Blog Article*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.excerpt
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  excerpt: prismic.RichTextField;

  /**
   * Featured Image field in *Blog Article*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.featured_image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  featured_image: prismic.ImageField<never>;

  /**
   * Category field in *Blog Article*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.category
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  category: prismic.ContentRelationshipField<"blog_category">;

  /**
   * Slice Zone field in *Blog Article*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#slices
   */
  slices: prismic.SliceZone<BlogArticleDocumentDataSlicesSlice> /**
   * Meta Title field in *Blog Article*
   *
   * - **Field Type**: Text
   * - **Placeholder**: My title for Search Engine
   * - **API ID Path**: blog_article.meta_title
   * - **Tab**: SEO and Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Blog Article*
   *
   * - **Field Type**: Text
   * - **Placeholder**: The description that will appear in search engine
   * - **API ID Path**: blog_article.meta_description
   * - **Tab**: SEO and Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_description: prismic.KeyTextField;

  /**
   * CanonicalLink field in *Blog Article*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.canonicalLink
   * - **Tab**: SEO and Metadata
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  canonicalLink: prismic.LinkField;

  /**
   * Social Cards - Facebook & Twitter field in *Blog Article*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_article.social_cards[]
   * - **Tab**: SEO and Metadata
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  social_cards: prismic.GroupField<
    Simplify<BlogArticleDocumentDataSocialCardsItem>
  >;
}

/**
 * Blog Article document from Prismic
 *
 * - **API ID**: `blog_article`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type BlogArticleDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<BlogArticleDocumentData>,
    "blog_article",
    Lang
  >;